Ask Claude to implement pagination for list interfaces.
Prompt: Currently the `listClients()` and `listUserGrants()` functions won't work correctly if there are more than 1000 results, because Workers KV only returns a maximum of 1000. We should support pagination in the same way Workers KV does. So, both of these functions should support an options struct containing `limit` and `cursor`, which are passed down to the Workers KV `list()` operation. They should also return a structure which contains a `cursor` if and only if there are more results.

Claude messed up here! For some reason it moved the `export class OAuthProvider {` line to the top of the file, which makes the whole thing not parse.