Ask Claude to fix its bugs.
Prompt: Oops, you accidentally moved the `export class OAuthProvider {` line to the top of the file, now it doesn't parse correctly.

Claude responded to this by making the code even more messed up, so I asked it to try again.

Prompt: No, this last edit made things worse. Let's back up and try again. The only thing you need to do is move the text `export class OAuthProvider {` and the associated doc comment back down so that it appears immediately after the declaration of the ListResult interface. Please don't re-order any of the other declarations.

This time Claude fixed the problem, but:
- It also re-ordered some declarations, even though I told it not to. AFAICT they aren't changed, just reordered, and it also added some doc comments.
- It fixed an unrelated bug, which is that `getClient()` was marked `private` in `OAuthProvider` but was being called from inside `OAuthHelpers`. I hadn't noticed this before, but it was indeed a bug.

Oh well, let's accept it and move on.