Ask Claude to allow the access token lookup and grant lookup to run in parallel.
Prompt: It would be nice if we could look up the token record and the grant record in parallel when authorizing a request. But to do that, we need to encode the user ID and the grant ID into the token itself, so that we have enough information to look up the grant before we've looked up the token. Maybe the full bearer token format can be `{userId}:{grantId}:{random-secret}`. Note that the token ID should still be a hash of the entire token including these new components, not just the secret part.