Ask Claude to move `props` into `ctx.props`.
Prompt: I'd like to make `defaultHandler` and `apiHandler` look more like standard Cloudflare Workers handlers, so that OAuthProivder looks more like a pure wrapper. To that end, I'd like to start with the following change: Instead of passing `props` as a parameter to `defaultHandler`, let's set it on the `ctx` object, that is, assign `ctx.props` to be the `props` value. This may sound a bit weird, but it actually aligns with an upcoming feature of Cloudflare Workers where `ctx.props` is used to communicate exactly this kind of information.

Claude first tried to update `DefaultHandler` even thought it doesn't have `props`.

Additional prompt: `props` is only passed to `ApiHandler`, so you don't need to update `DefaultHandler`. Also, please assume `ExecutionContext` already defines the property `props?: any` -- this is actually already the case in the latest version of Cloudflare Workers.

Additional prompt: Can you also update the sample code in the README?