Manually clean up that last readme change a bit.
I was impressed that Claude decided to make one of the two handlers be a WorkerEntrypoint while keeping the other one an object, to illustrate both. However, I think having the API handler be the WorkerEntrypoint works better. I also didn't like the duplication Claude produced.

I was a little tired of prompting so fixed this myself.