Manually switch key-wrapping format from 'jwk' to 'raw'.
AES-KW is designed to operate on raw keys. Giving it 'jwk's tends to throw errors about the length not being a clean multiple of 64 bits.

When Claude was prompted to fix it, it decided the best solution was to switch away from AES-KW and use AES-GCM instead, in order to wrap an arbitrary key string. This made the code quite a bit more complicated and made the wrapped keys larger.

If we just use 'raw' format here, we solve the problem, and get to stick with AES-KW as it was intended to be used.