Ask Claude to add support for scopes_supported from a new constructor option.
I really could have done this one faster by hand, oh well.

prompt: I noticed scopes_supported in the RFC 8414 metadata is set to an empty array. We should add a new constructor option to `OAuthProvider` called `scopesSupported` so that the application can pass the right list in. This should be optional: if the application doesn't specify it, then we should omit `scopes_supported` from the metadata entirely.

prompt: This can be simpler: JSON.stringify() will already omit fields that are undefined, so you can just set `scopes_supported` to `this.options.scopesSupported` and it'll automatically be omitted from the JSON if not set.

prompt: Can you please update the readme?

Weirdly Claude tried to _commit_ the readme change, which I hadn't seen before, but I called it off.